Added support for an incremental project builder (aka a Builder). The current builder just prints log messages when it is invoked, but this will later invoke the PTJava compiler. The builder is added to the current project through a Nature.
The nature for the builder (PTJavaBuilderNature) can be associated with a 
project by the menu action in the PTJava menu on the workbench menu bar. This
will later be replaced by a Custom project option.

Will now work on invoking the compiler when a .ptjava file is saved in the
project.