Massive changes have been made to the text editor:
The text editor is now a subclass of the JDT text editor (in particular,
PTJavaTextEditor is now a subclass of CompilationUnitEditor). This means
that all the features of the JDT editor are now a part of the PTJava 
editor (at the expense of a few warnings and accessing classes that are
not designed to be accessed in org.eclipse.jdt). This includes text hovers,
outline view, syntax highlighting, code completion. The PTJava compiler 
still builds ptjava files on request too.

The most prominent changes have occured in PTJavaSourceViewerConfiguration,
which now subclasses JavaSourceViewerConfiguration and uses code scanners/
preferences from there (and adding extra scanners, etc when required).

Subclasses have been updated to take on preferences of the JDT editor,
including syntax highlight colors and bracket completion.

To Do: 

Clean up code to remove warnings.

Integrate plug-in with a "Create PTJava Project" interface.

Update documentation.