Added basic data structures for TASKs and began setting up Table Viewer for showing these tasks. Will complete the view at a later stage.
Have also investigated the outcome of using the JDT ASTParser on a ptjava file.

Given the following:

public class Foo {
	TASK public Integer task1() {
		return new Integer(0);
	}
	TASK public Integer task2() {
		return new Integer(1);
	}
	public void standardMethod() {
		TaskID<Integer> id1= task1();
		TaskID<Integer> id2= task2() dependsOn(id1);
	}
}

The ASTParser will parse standardMethod as a method, but the "dependsOn" clause
will cause the generation of method invocation nodes to fail. i.e. the AST will
look like the following if the methods are printed:

	public Integer task1() {
		return new Integer(0);
	}
	public Integer task2() {
		return new Integer(1);
	}
	public void standardMethod() {
	}