Updated plug-in to work with new compiler and runtime.
Compiler and runtime classes were renamed, so had to fix build and classpath 
issues.

Updated documentation to help Users that wish to generate their own compiler and
runtime jars.

Fixed a bug in the preferences page (whereby the "Use Defaults" button did not
disable the FileFieldEditor's path)

Rebuilt the PTCompiler.jar and PTRuntime.jar for plug-in and tested it.